Added .env.local and message.txt check exists(not overwriting it)

+++ install.sh
@@ -7,10 +7,15 @@ sudo cp telegram-send.sh /usr/bin/telegram-send
 sudo chown root:root /usr/bin/telegram-send
 
 mkdir -p /var/serveraccesslogging
-sudo cp .env /var/serveraccesslogging/.env.local
-sudo cp message.txt /var/serveraccesslogging/message.txt
+if test -f "/var/serveraccesslogging/.env.local"; then
+  sudo cp .env /var/serveraccesslogging/.env.local
+fi
 
-#Copying the notify script into the profile.d folder to make it execute on login
+if test -f "/var/serveraccesslogging/message.txt"; then
+  sudo cp message.txt /var/serveraccesslogging/message.txt
+fi
+
+#Copying notify script into the profile.d folder to make it execute on login
 sudo cp access-notify.sh /etc/profile.d/access-notify.sh
 
 echo "Completed! Now the file .env is in this folder: /var/serveraccesslogging/.env, and you can customize the message in the same folder but in the file message.txt"
